Convert ring buffer into template class
