+++ sysroot/usr/include/kernel/lib/ring_buffer.hpp
@@ -13,80 +13,137 @@
 #pragma once
 
 #include <stdint.h>
+#include <lib/errno.h>
 
-class RingBuffer {
+template <typename T> class RingBuffer {
 public:
     /**
      * @brief Initializes the circular buffer and allocates
      * the data memory.
      *
      * @param size Maximum buffer capacity (in bytes)
-     * @param data Data buffer to be used. If NULL, a buffer of
+     * @param buf Data buffer to be used. If NULL, a buffer of
      * size will be allocated and internally managed. If a buffer
      * is provided, the caller is responsible for memory management
      * of the provided buffer.
      * @return int Returns 0 on success and -1 on error.
      */
-    RingBuffer(int size, uint8_t* buf);
-    /**
-     * @brief Destroys a circular buffer and frees all the
-     * data memory.
-     */
-    ~RingBuffer();
+    explicit RingBuffer(int size, T* buf)
+        : head(0)
+        , tail(0)
+        , length(0)
+        , capacity(size)
+    {
+        this->data = buf;
+    }
     /**
      * @brief Writes a byte into the circular buffer.
      *
-     * @param byte Byte to write to the buffer
+     * @param val Data to write to the buffer
      */
-    int Enqueue(uint8_t byte);
+    int Enqueue(T val)
+    {
+        int status = 0;
+        // Check if the buffer is full. If so, we can't enqueue.
+        if (IsFull()) {
+            status = -1;
+            errno = ENOBUFS;
+        } else {
+            // Write the data at the write index
+            this->data[this->head] = val;
+            this->head = ((this->head + 1) % this->capacity);
+            ++this->length;
+        }
+        // Return the status code
+        return status;
+    }
     /**
      * @brief Removes a bytes from the circular buffer and writes
      * the value of the byte to the data pointer.
      *
-     * @param buf Pointer to uint8_t to contain the data
+     * @param buf Buffer to contain the data
      * @return int Returns 0 on success and -1 on error.
      */
-    int Dequeue(uint8_t* buf);
+    int Dequeue(T* buf)
+    {
+        int status = 0;
+        // Check if the buffer is empty. If so, we can't dequeue
+        if (IsEmpty()) {
+            status = -1;
+            errno = EINVAL;
+        } else {
+            // Read out the data and decrement the position
+            *buf = this->data[this->tail];
+            this->tail = ((this->tail + 1) % this->capacity);
+            --this->length;
+        }
+        // Return the status code
+        return status;
+    }
     /**
      * @brief Grab the latest bytes of data from the buffer without
      * removing it.
      *
-     * @param buf Pointer to uint8_t to contain the data
+     * @param buf Buffer to contain the data
      * @return int Returns 0 on success and -1 on error.
      */
-    int Peek(uint8_t* buf);
+    int Peek(T* buf)
+    {
+        int status = 0;
+        // Check if the buffer is empty. If so, we can't dequeue
+        if (IsEmpty()) {
+            status = -1;
+            errno = EINVAL;
+        } else {
+            // Read out the data and don't decrement the position
+            *buf = this->data[this->tail];
+        }
+        // Return the status code
+        return status;
+    }
     /**
      * @brief Query whether the circular buffer is empty.
      *
      * @return true The buffer is empty
      * @return false The buffer is not empty
      */
-    bool IsEmpty();
+    bool IsEmpty()
+    {
+        return (this->length == 0);
+    }
     /**
      * @brief Query whether the circular buffer is full.
      *
      * @return true The buffer is full
      * @return false The buffer is not full
      */
-    bool IsFull();
+    bool IsFull()
+    {
+        return (this->length == this->capacity);
+    }
     /**
      * @brief Returns the number of items (bytes) in the buffer.
      *
      * @return int Number of bytes available for reading.
      */
-    int Length();
+    int Length()
+    {
+        return this->length;
+    }
     /**
      * @brief Returns the buffer capacity (in number of bytes).
      *
      * @return int Buffer capacity in bytes
      */
-    int Capacity();
+    int Capacity()
+    {
+        return this->capacity;
+    }
 
 private:
-    uint8_t* data;
+    T* data;
     int head;
     int tail;
     int length;
     int capacity;
-    bool external;
 };
